get_famd {factoextra}	R Documentation

Extract the results for individuals and variables - FAMD

Description

Extract all the results (coordinates, squared cosine and contributions) for the active individuals and variables from Factor Analysis of Mixed Date (FAMD) outputs.

get_famd(): Extract the results for variables and individuals

get_famd_ind(): Extract the results for individuals only

get_famd_var(): Extract the results for quantitative and qualitative variables only

Usage

get_famd(res.famd, element = c("ind", "var", "quanti.var", "quali.var"))

get_famd_ind(res.famd)

get_famd_var(res.famd, element = c("var", "quanti.var", "quali.var"))

Arguments

res.famd
an object of class FAMD [FactoMineR].

element
the element to subset from the output. Possible values are "ind", "quanti.var" or "quali.var".

Value

a list of matrices containing the results for the active individuals and variables, including :

coord
coordinates of indiiduals/variables.

cos2
cos2 values representing the quality of representation on the factor map.

contrib
contributions of individuals / variables to the principal components.

Author(s)

Alboukadel Kassambara alboukadel.kassambara@gmail.com

Examples

 # Compute FAMD
 library("FactoMineR")
 data(wine)
 res.famd <- FAMD(wine[,c(1,2, 16, 22, 29, 28, 30,31)], graph = FALSE)
 
 # Extract the results for qualitative variable categories
 quali.var <- get_famd_var(res.famd, "quali.var")
 print(quali.var)
 head(quali.var$coord) # coordinates of qualitative variables
 
 # Extract the results for quantitative variables
 quanti.var <- get_famd_var(res.famd, "quanti.var")
 print(quanti.var)
 head(quanti.var$coord) # coordinates
 
 # Extract the results for individuals
 ind <- get_famd_ind(res.famd)
 print(ind)
 head(ind$coord) # coordinates of individuals
 
 

[Package factoextra version 1.0.6 Index]